cli-secure-messenger/requirements.txt
websockets==12.0
PyNaCl==1.5.0
argon2-cffi==23.1.0
